ENH: Instance editor example working.
Each model object is now registered in a registry, and the set_trait
/ get_trait function looks up in the registry for a given model id to
hook up trait change events. This enables us to reuse an instance's
html/js code to provide instance editor's html and js.

Also, id of a model object maps to a data attribute in html, not id,
since the same model can have multiple views in a session.