WIP: examples work with a real async web backend.
The backend uses websockets entirely.  The notebook example now works
with any notebook server.  The web tests do not pass yet.  There are
also issues on the JS side when angular is not in the picture as the
attributes are typically undefined or null when asked for and then are
magically set.  We need a nice way to deal with this.