Send object info while marshaling
This makes the unmarshal step completely synchronous and also makes the
`create_proxy` method synchronous, requiring no server requests for
object information. This requires that we send the object information
when the object is marshaled in the Python side.  As a result the
implementation is much cleaner, the jigna JS client API is leaner as we
do not need get_list/dict/instance_info methods at all.