Refactor async functionality into AsyncClient.
This cleans up the code a lot and localizes the async related changes.
This did require some cleanup of the current client which was doing a
few things in the constructor which it shouldn't have.  Also did some
cleanup to avoid repeating code.  The way in which we turn on async in
the tests still needs to be fixed and this will be done later.