Singificantly improve performance.
The async tests also seem to work now.  The performance improvement is
by not sending all the attribute information, instead we send out
default attribute information at first which is stored and reused.  The
async client returns the defaults if the cache is invalid thus
preventing template rendering failures.  This is still slower than the
original approach where only the length is sent but async works well now.