Fix a few issues.
Pull the code that returns the default value into the instance attribute
getter instead of the client.  Don't set the cache when the default is
returned.  Execute the js again to fix the failing tests -- this is
because with the async backend one may get old data which is not
undefined, so one must get the data again.