BUG: Fix typo. Better default types for list/dict.
Fixes a silly cut/paste typo in proxy_factory's instance attribute
getter.  Earlier, the __info__ default type would potentially have a
large default list which is rather wasteful.  For example if the first
time a type is encountered and it has a huge number of elements in a
list/dict, the __info__ attribute would contain that and consequently
any new instance would automatically appear as if it had many elements.
Now it only has an empty list/dict which is filled up on demand.