Refactor to cleanup.
This version cleans up the async support.  The `server.py` is no longer
changed, instead the `web_server.py` has the additional async related
option and code.  The JS layer is also much cleaner and the async
related proxy code is in a new `async_proxy_factory.js` which is used
when the `AsyncClient` is used.  The original sync version is not
changed much at all.