Add a new event to create any new types.
When one has a large number of nested objects, the new type info should
be sent everytime a new type is encountered.  This is important in the
async mode as new types may be dynamically encountered and the JS side
should have a constructor ready when an instance is needed.  The event
elegantly allows for this to work.  A JS listener simply creates a new
constructor when new type information is available.