Fix bug with repeated list event.
When a list is changed it so happens that sometimes two identical
websocket messages are received on the JS side even though only one
event was sent from the Python side.  This causes JS proxy lists to go
out of sync with the Python lists and acquire copies of existing
elements.  This is fixed by only handling list events on the JS side if
the length of the proxy matches that on the Python side.

Also switch to using the Chrome driver on Darwin.  Newer versions of
Firefox (>45) do not work with selenium.